+++ example/web/stub/CastButtonStub.tsx
@@ -1,4 +1,4 @@
-import type { ViewProps } from 'react-native';
+import type {ViewProps} from 'react-native';
 import React from 'react';
 
 // Provides a stub for react-native-google-cast on web.