+++ src/ui/components/button/SkipButton.tsx
@@ -65,8 +65,8 @@ export class SkipButton extends PureComponent<SkipButtonProps, SkipButtonState> 
 
   private readonly onPlaying = () => {
     const player = (this.context as UiContext).player;
-    const isCasting =  player.cast.chromecast?.casting ?? false
-    this.setState({ enabled: player.seekable.length > 0 || player.buffered.length > 0 || isCasting});
+    const isCasting = player.cast.chromecast?.casting ?? false;
+    this.setState({ enabled: player.seekable.length > 0 || player.buffered.length > 0 || isCasting });
   };
 
   private readonly onPress = () => {