+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fixed an issue where the `SeekBar`'s seekable state was not updated when switching to a MP4 source.
+- Fixed an issue where the `SkipButton` components are not rendered when switching sources while casting.
 
 ## 0.7.1 (2024-04-16)