+++ CHANGELOG.md
@@ -10,6 +10,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.8.0 (2024-09-11)
+
+### Added
+
+- Added support for THEOplayer 8.0.
+
+### Fixed
+
+- Fixed a deprecation warning by removing `defaultProps`.
+
 ## 0.7.2 (2024-05-15)
 
 ### Fixed